Add a recap of player's stats change to update
Fix loadplayer() and updateplayers() to correctly store player's
name and kingdom.
Rework playimetolong() to not lose *playtime.
Rework playtimetostr() to handle plural correctly.
Add some error handling to dlimg.

+++ util.h
@@ -1,4 +1,4 @@
-#define LENGTH(X)   (sizeof X / sizeof X[0])
+#define LENGTH(X) (int)(sizeof X / sizeof X[0])
 
 void die(const char *errstr, ...);
 char *nstrchr(const char *s, int c, int n);
